pyproject.toml
[project]
name = "mybro"
version = "0.1.0"
description = "Local project manager + universal ticketing system for Claude Code"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.20.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.78.0",
    "pynput>=1.7.6",
    "structlog>=24.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pyobjc-framework-Quartz>=10.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.setuptools.packages.find]
include = ["backend*", "tracker*", "tpm*"]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"
